build: move shared.lua to _core/
